implement multi_subscriber for Thread
